Task3
Task3

+++ Assets/Script/System/ProgressBar.cs
@@ -8,13 +8,14 @@ public class ProgressBar : MonoBehaviour
 
     public Slider slider;
 
-    public void SetMaxProgress(int progress)
+    public void SetMaxProgress(float progress)
     {
         slider.value = progress;
         slider.maxValue = progress;
     }
-   public void SetProgress(int progress)
+   public void SetProgress(float progress)
     {
+        
         slider.value = progress;
     }
 }
